исправление точек маршрута
